requirements.txt
docker
pyyaml
lxml
